requirements.txt
google-genai
python-telegram-bot[job-queue]
flask
supabase
google-auth-oauthlib
google-api-python-client
pytz